# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Darbar Multi-Cuisine Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specials:
    get:
      operationId: listSpecials
      tags: [specials]
      summary: Get today's special offers
      responses:
        "200":
          description: Today's specials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Special"
    post:
      operationId: createSpecial
      tags: [specials]
      summary: Create a special offer (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Special"
...
  /specials/{id}:
    patch:
      operationId: updateSpecial
      tags: [specials]
      summary: Update a special offer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Special"
    delete:
      operationId: deleteSpecial
      tags: [specials]
      summary: Delete a special offer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List approved customer reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a customer review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /admin/reviews/{id}:
    patch:
      operationId: updateReviewApproval
      tags: [reviews]
      summary: Approve or reject a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApprovalUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
...
components:
  schemas:
...
    Special:
      type: object
      properties:
        id:
          type: integer
        dishName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - dishName
        - price
        - isActive
        - createdAt
    SpecialInput:
      type: object
      properties:
        dishName:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        imageUrl:
          type: string
        isActive:
          type: boolean
      required:
        - dishName
        - price
    SpecialUpdate:
      type: object
      properties:
        dishName:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        imageUrl:
          type: string
        isActive:
          type: boolean
    Review:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - customerName
        - rating
        - comment
        - isApproved
        - createdAt
    ReviewApprovalUpdate:
      type: object
      properties:
        isApproved:
          type: boolean
      required:
        - isApproved
    ReviewInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 1
      required:
        - customerName
        - rating
        - comment
...